moved content types to config file
